refactor: Refactored repeated elements to components for reusability
Created components for repeated elements found in dashboard pages: sidebar, doc
links, cards. Makes it easier to repeat elements and change component info using
props. It cleans up code making it easier to read and understand.
