+++ pages/dashboard/index.tsx
@@ -9,6 +9,7 @@ import ArrowForwardIosIcon from '@material-ui/icons/ArrowForwardIos';
 import AnnouncementCard from './Components/AnnouncementCards';
 import MentorCard1 from './Components/MentorCard1';
 import MentorCard3 from './Components/MentorCard3';
+import Sidebar from './Components/Sidebar';
 
 /**
  * The dashboard / hack center.
@@ -40,19 +41,7 @@ export default function Dashboard() {
         <meta name="description" content="HackPortal's Dashboard" />
       </Head>
 
-      {/* ghost section to fill in for fixed sidebar */}
-      <section
-        id="ghost"
-        className="flex justify-center h-screen sticky top-0 lg:w-1/8 md:w-1/7 w-1/6"
-      ></section>
-
-      <section
-        id="Sidebar"
-        className="flex justify-center items-center h-screen fixed top-16 border-r-2 border-gray-600 lg:w-1/8 md:w-1/7 w-1/6 text-xs lg:text-sm text-center"
-      >
-        <div>Welcome, {!user || !isSignedIn ? 'hacker' : user.firstName}</div>
-        <div className="text-indigo-500">{role}</div>
-      </section>
+      <Sidebar />
 
       <section id="mainContent" className="px-6 py-3 w-5/6 lg:wd-7/8 md:w-6/7">
         <section id="subheader" className="w-full pb-6 sticky top-16">
